feat(models): Update models

+++ shared/models/data-policy.go
@@ -1,10 +1,16 @@
 package models
 
 type DataPolicy struct {
-	Id                   int    `json:"id" validate:"-"`
-	Descr                string `json:"descr" validate:"required,min=2,max=255"`
-	UseAggregation       bool   `json:"use-aggregation" validate:"-"`
-	Retention            int    `json:"retention" validate:"required,min=1"`             // Raw data retantion in hours.
-	AggregationRetention int    `json:"aggregation-retention" validate:"required,min=1"` // Aggregated data retation in hours.
-	AggregationInterval  int    `json:"aggregation-interval" validate:"required,min=1"`  // Aggregated data interval in seconds.
+	// Id is the identifier.
+	Id int16 `json:"id" validate:"-"`
+	// Descr is the data policy description.
+	Descr string `json:"descr" validate:"required,min=2,max=255"`
+	// UseAggregation is the aggregation enabled status.
+	UseAggregation bool `json:"use-aggregation" validate:"-"`
+	// Retention is the raw data retention in hours.
+	Retention int32 `json:"retention" validate:"required,min=1"`
+	// AggregationRetention is the aggregation retention in hours.
+	AggregationRetention int32 `json:"aggregation-retention" validate:"required,min=1"`
+	// AggregationInterval is the aggregation interval in seconds.
+	AggregationInterval int32 `json:"aggregation-interval" validate:"required,min=1"`
 }
