+++ shared/models/rts.go
@@ -2,9 +2,9 @@ package models
 
 type RTSMetricInfo struct {
 	// PullingTimes is how many times will pull the data.
-	PullingTimes int
+	PullingTimes int16
 	// PullingInterval is the interval between each data request in miliseconds.
-	PullingInterval int
+	PullingInterval int32
 	// Cache duration is the cached data durantion in miliseconds.
-	CacheDuration int
+	CacheDuration int32
 }
