+++ shared/models/snmp.go
@@ -2,16 +2,16 @@ package models
 
 type SNMPContainer struct {
 	// ContainerId is the container id.
-	ContainerId int `json:"container-id" validate:"-"`
+	ContainerId int32 `json:"container-id" validate:"-"`
 
 	// CacheDuration is the cache duration in miliseconds of this configuration on the SNMP service.
-	CacheDuration int `json:"cache-duration" validate:"required"`
+	CacheDuration int32 `json:"cache-duration" validate:"required"`
 
 	// Target is an ipv4 address.
 	Target string `json:"target" validate:"required,max=50"`
 
 	// Port is a port.
-	Port uint16 `json:"port" validate:"required,max=65535"`
+	Port int32 `json:"port" validate:"required,max=65535"`
 
 	// Transport is the transport protocol to use ("udp" or "tcp"); if unset "udp" will be used.
 	Transport string `json:"transport" validate:"required,max=3"`
@@ -20,10 +20,10 @@ type SNMPContainer struct {
 	Community string `json:"community" validate:"required,max=50"`
 
 	// Timeout is the timeout for one SNMP request/response.
-	Timeout int `json:"timeout" validate:"required,max=60000"`
+	Timeout int32 `json:"timeout" validate:"required,max=60000"`
 
 	// Set the number of retries to attempt.
-	Retries int `json:"retries" validate:"required"`
+	Retries int16 `json:"retries" validate:"required"`
 
 	// MsgFlags is an SNMPV3 MsgFlags.
 	MsgFlags uint8 `json:"msg-flags" validate:"-"`
@@ -32,12 +32,12 @@ type SNMPContainer struct {
 	Version uint8 `json:"version" validate:"min=0,max=3"`
 
 	// Max oids per request.
-	MaxOids int `json:"max-oids" validate:"required"`
+	MaxOids int16 `json:"max-oids" validate:"required"`
 }
 
 type SNMPMetric struct {
 	// MetricId is the metric identifier.
-	MetricId int `json:"-" validate:"-"`
+	MetricId int64 `json:"-" validate:"-"`
 	// OID is the snmp object identifier.
 	OID string `json:"oid" validate:"required,max=128"`
 }
